:recycle: refactor(resume): change order of resume and modify education
